restructure linear regression classes
