+++ profit/sur/linreg/__init__.py
@@ -1,0 +1,4 @@
+from .linear_regression import LinearRegression
+from .chaospy_linreg import ChaospyLinReg
+from .sklearn_linreg import SklearnLinReg
+from .custom_linreg import CustomLinReg
